+++ terraform/bootstrap/providers.tf
@@ -1,0 +1,4 @@
+provider "aws" {
+  region              = var.aws_region
+  allowed_account_ids = ["390017876013"]
+}